% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_functions.R
\name{build_eitc_lookup}
\alias{build_eitc_lookup}
\title{Build EITC Lookup Table}
\usage{
build_eitc_lookup(eitc_params)
}
\arguments{
\item{eitc_params}{Dataframe with EITC parameters from tax_fed_credits_df}
}
\value{
Dataframe with EITC parameters by eitc_children and household_type
}
\description{
Creates a lookup table with EITC parameters for all combinations of
number of children (0-2) and household types
}
