% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_functions.R
\name{extract_cdctc_params}
\alias{extract_cdctc_params}
\title{Extract CDCTC Parameters}
\usage{
extract_cdctc_params(cdctc_params)
}
\arguments{
\item{cdctc_params}{Dataframe with CDCTC parameters from tax_fed_credits_df}
}
\value{
Named list of CDCTC parameters
}
\description{
Extracts Child and Dependent Care Tax Credit parameters from tax data
}
