% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_functions.R
\name{extract_ctc_params}
\alias{extract_ctc_params}
\title{Extract CTC Parameters}
\usage{
extract_ctc_params(ctc_params)
}
\arguments{
\item{ctc_params}{Dataframe with CTC parameters from tax_fed_credits_df}
}
\value{
Named list of CTC parameters
}
\description{
Extracts Child Tax Credit parameters from tax data
}
