% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_functions.R
\name{load_fed_payroll_parameters}
\alias{load_fed_payroll_parameters}
\title{Load Federal Payroll Tax Parameters}
\usage{
load_fed_payroll_parameters(tax_fed_payroll_df, year)
}
\arguments{
\item{tax_fed_payroll_df}{Dataframe with federal payroll tax parameters}

\item{year}{Tax year}
}
\value{
Named list of payroll tax parameters
}
\description{
Extracts federal payroll tax parameters for a specific year
}
