% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print_convergence_summary}
\alias{print_convergence_summary}
\title{Print Convergence Summary}
\usage{
print_convergence_summary(df, debug = FALSE)
}
\arguments{
\item{df}{Result dataframe with converged column}

\item{debug}{If TRUE, show additional details}
}
\description{
Displays convergence statistics after solver completes
}
