% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_functions.R
\name{calculate_federal_payroll_taxes}
\alias{calculate_federal_payroll_taxes}
\title{Calculate Federal Payroll Taxes}
\usage{
calculate_federal_payroll_taxes(calculations_df, tax_fed_payroll_df, year)
}
\arguments{
\item{calculations_df}{Dataframe with starting_income and household_type}

\item{tax_fed_payroll_df}{Dataframe with federal payroll tax parameters}

\item{year}{Tax year}
}
\value{
Dataframe with ss_tax, medicare_tax, and total_fed_payroll_tax columns added
}
\description{
Calculates Social Security and Medicare taxes
}
